+++ Examples/Console/AI/example_AI_on_demand_in_loop.cs
@@ -4,9 +4,30 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-class example_AI_on_demand_once
-{
-    static void Main1()
+
+// Question 
+    // 1. Multiple main in one project
+    // 2. Static function and thread?
+    // 3. Doxygen
+    // 4. Add .cs file into doxygen
+    // 5. WifiDAQE3A replace with all type  
+
+class example_AI_on_demand_once_in_loop
+{ 
+    static void loop_func(WifiDAQE3A handle, int port, int delay , int timeout =3) 
+    {
+        int t = 0;
+        while (t < timeout) 
+        {  
+            // Set AI port to 1 and acquisition mode to on demand mode (0)
+            List<float> ch = handle.AI_readOnDemand(port);
+            Console.WriteLine($"data: {ch[0]}, {ch[1]}, {ch[2]}, {ch[3]}, {ch[4]}, {ch[5]}, {ch[6]}, {ch[7]},");
+            t += delay; 
+        }
+        Console.WriteLine("loop_func end");
+    }
+
+    static void Main()
     {
         Console.WriteLine("Start example code...");
 
@@ -17,7 +38,6 @@ class example_AI_on_demand_once
         Console.WriteLine($"{dev.getDriverName()} - Version {dev.getDriverVersion()}");
 
         // Connect to network device
-
         dev.connect("192.168.5.79");
 
         // Perform DAQ basic information 
@@ -27,7 +47,7 @@ class example_AI_on_demand_once
             int status;
             int port = 1;
             int mode = 0;
-
+             
             // Get firmware model & version
             string[] driver_info = dev.Sys_getDriverInfo();
             Console.WriteLine($"Model name: {driver_info[0]}");
@@ -40,10 +60,12 @@ class example_AI_on_demand_once
             // Set AI port to 1 and acquisition mode to on demand mode (0)
             status = dev.AI_setMode(port, mode);
             Console.WriteLine($"AI_setMode status: {status}");
+            
+            //Thread ai_thread = new Thread(() => loop_func(dev, port, 1, 3));
+            //ai_thread.Start();
 
-            // Set AI port to 1 and data acquisition
-            List<float> ch = dev.AI_readOnDemand(port);
-            Console.WriteLine($"data: {ch[0]}, {ch[1]}, {ch[2]}, {ch[3]}, {ch[4]}, {ch[5]}, {ch[6]}, {ch[7]}");
+            loop_func(dev, port, 1, 3);
+            
 
             // Close port 1
             status = dev.AI_close(port);